+++ addon/CHANGELOG.md
@@ -1,4 +1,4 @@
-##### *GENERATED FILE [2026-07-31 11:14 UTC] — do not edit by hand.*
+##### *GENERATED FILE [2026-07-31 11:57 UTC] — do not edit by hand.*
 
 # Changelog